src/test/kotlin/day25/Day40SolverTest.kt
package day25

import helpers.readDataAsList
import io.kotlintest.shouldBe
import org.junit.jupiter.api.Test

internal class Day40SolverTest {
    private val daySolver = Day40Solver()
    private val dir = "day40"

    @Test
    fun `should return correct result for sample data for part1`() {
        val data = readDataAsList(dir, "sample")
        daySolver.y = 10
        daySolver.part1(data) shouldBe 26
    }

    @Test
    fun `should return correct result for original data for part1`() {
        val data = readDataAsList(dir, "original")
        daySolver.y = 2000000
        daySolver.part1(data) shouldBe 5716881L
    }

    @Test
    fun `should return correct result for sample data for part2`() {
        val data = readDataAsList(dir, "sample")
        daySolver.part2(data) shouldBe 52000002L
    }

    @Test
    fun `should return correct result for original data for part2`() {
        val data = readDataAsList(dir, "original")
        daySolver.part2(data) shouldBe 10852583132904
    }
}